+++ src/ATAP.Utilities.BuildTooling.DotnetBuild.PowerShell/tests/Unit/DotnetBuildChild.Contract.Tests.ps1
@@ -20,6 +20,10 @@ BeforeAll {
     'Compress-PSModuleArtifacts'
     'Get-BuildContext'
     'Get-PSModuleVersionFromNBGV'
+    'Install-DabGlobalTool'
+    'Initialize-DabMcpConfiguration'
+    'Initialize-DabMcpServer'
+    'Add-DabMcpEntity'
     'Invoke-DotnetBuildWithRetry'
     'Invoke-DotnetNuGetPush'
     'Invoke-ModuleBuildWithRetry'
@@ -29,6 +33,9 @@ BeforeAll {
     'Parse-MSBuildFile'
     'Resolve-FeatureSlug'
     'Resolve-PSModuleMetadata'
+    'Start-DabMcpServer'
+    'Test-DabInstallation'
+    'Test-DabMcpConfiguration'
   )
 
   Remove-Module 'ATAP.Utilities.BuildTooling.DotnetBuild.PowerShell' -Force -ErrorAction SilentlyContinue
@@ -43,7 +50,7 @@ Describe 'DotnetBuild child module contract' -Tag 'Unit', 'Contract' {
 
   It 'declares accepted immutable dependency floors' {
     $manifest = Import-PowerShellDataFile -LiteralPath $script:manifestPath
-    [string]$manifest.ModuleVersion | Should -BeExactly '0.1.1'
+    [string]$manifest.ModuleVersion | Should -BeExactly '0.1.2'
     $dependencies = @{}
     foreach ($requiredModule in $manifest.RequiredModules) {
       $dependencies[$requiredModule.ModuleName] = [string] $requiredModule.ModuleVersion